requirements.txt
# Minimal requirements for API-based RAG ChatBot
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0

# Lightweight text processing (no heavy ML libraries)
scikit-learn==1.3.2
numpy==1.24.3

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0

# HTTP requests for API calls
requests==2.31.0

# CORS support
starlette==0.27.0
